% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{condense.to.other.all}
\alias{condense.to.other.all}
\title{condense.to.other.all}
\usage{
condense.to.other.all(df)
}
\description{
Condense all categories with less than 10% of frequency into an 'Others' category
}
