% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{condense.to.other}
\alias{condense.to.other}
\title{condense.to.other}
\usage{
condense.to.other(df, x)
}
\description{
Condense specific category with less than 10% of frequency into an 'Others' category
}
