% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{mean.bar.plot}
\alias{mean.bar.plot}
\title{mean.bar.plot}
\usage{
\method{mean}{bar.plot}(dataset, x, y, z)
}
\description{
Creates a barplot representing the mean, standard deviations
and jitter points
}
